+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The subsection are to be interpreted as defined [here](http://keepachangelog.com
 - Updated gitignore.
 - Linted and style checked all the code.
 - `src/Document.js` cleaned up.
+- Shortened `COMMAND` to `CMD`.
 
 ## [0.1.6] - 2015-03-19